Created new version based on baseapp
