+++ TeaTimer/app/src/main/java/com/jamescoggan/teatimer/di/components/GitHubComponent.java
@@ -1,0 +1,13 @@
+package com.jamescoggan.teatimer.di.components;
+
+import com.jamescoggan.teatimer.di.modules.GitHubModule;
+import com.jamescoggan.teatimer.di.scopes.UserScope;
+import com.jamescoggan.teatimer.fragments.HomeFragment;
+
+import dagger.Component;
+
+@UserScope
+@Component(dependencies = NetComponent.class, modules = GitHubModule.class)
+public interface GitHubComponent {
+    void inject(HomeFragment homeFragment);
+}
