+++ TeaTimer/app/src/main/java/com/jamescoggan/teatimer/models/Repository.java
@@ -1,4 +1,4 @@
-package com.jamescoggan.baseapp.models;
+package com.jamescoggan.teatimer.models;
 
 import io.realm.RealmObject;
 import io.realm.annotations.PrimaryKey;
