+++ TeaTimer/app/src/main/res/layout/fragment_home.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.jamescoggan.baseapp.fragments.HomeFragment">
+    tools:context="com.jamescoggan.teatimer.fragments.HomeFragment">
 
     <android.support.design.widget.FloatingActionButton
         android:id="@+id/home_fab"
